rename app js
